+++ mod-record-specifications-server/src/test/java/org/folio/api/SpecificationStorageSyncApiIT.java
@@ -19,7 +19,7 @@ import org.folio.spring.testing.extension.DatabaseCleanup;
 import org.folio.spring.testing.extension.EnableOkapi;
 import org.folio.spring.testing.extension.impl.OkapiConfiguration;
 import org.folio.spring.testing.type.IntegrationTest;
-import org.folio.support.IntegrationTestBase;
+import org.folio.support.SpecificationITBase;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
@@ -29,7 +29,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 @EnableOkapi
 @IntegrationTest
 @DatabaseCleanup(tables = FIELD_TABLE_NAME, tenants = TENANT_ID)
-class SpecificationStorageSyncApiIT extends IntegrationTestBase {
+class SpecificationStorageSyncApiIT extends SpecificationITBase {
 
   protected static OkapiConfiguration okapi;
 
@@ -99,6 +99,8 @@ class SpecificationStorageSyncApiIT extends IntegrationTestBase {
       .hasSize(1193)
       .extracting(UuidPersistable::getId)
       .containsExactlyInAnyOrder(createdIndicatorCodeIds);
+
+    assertSpecificationUpdatedEvent();
   }
 
   private UUID @NotNull [] toIdArray(List<? extends UuidPersistable> all) {